requirements.txt
python-dotenv
langchain-google-genai
langgraph
langgraph-supervisor
langmem
unstructured-client
langchain-unstructured
langchain-experimental
langchain-chroma
pandas
tabulate
streamlit